backend/requirements.txt
 
grpcio
grpcio-tools
